Renamed utils to autodiscovery
