+++ src/pushcart/stages/__init__.py
@@ -1,3 +1,3 @@
-from .utils import get_stage_object
+from .autodiscovery import get_stage_object
 
 __all__ = ["get_stage_object"]
